-- Added support for compling code on the Mac  -- For Java side, added API to get at a file's modification time  -- Fixed a bug in the re-replication code on the chunkserver: the     re-replicator was issuing writes that were not a multiple of     the checksum block size.  -- Fixed a bug with KfsPing code where it was only reading 1K of     response from the server.  With lots of servers, the response size     can get bigger.

+++ trunk/src/cc/libkfsClient/KfsRead.cc
@@ -363,7 +363,7 @@ KfsClient::DoLargeReadFromServer(int fd, char *buf, size_t numBytes)
 
     assert(chunk->chunkSize - pos->chunkOffset >= 0);
 
-    size_t numAvail = min(chunk->chunkSize - pos->chunkOffset, numBytes);
+    size_t numAvail = min((size_t) (chunk->chunkSize - pos->chunkOffset), numBytes);
     size_t numRead = 0;
 
     while (numRead < numAvail) {
